+++ SayHiSemanticUi/Readme.md
@@ -2,6 +2,7 @@
 <hr>
 
 - implemented apollo custom errors
+- created files&folders structure
 
 ### Installation
 - npm install apollo-server grahql mongoose
