+++ SayHiSemanticUi/graphql/resolvers/users.js
@@ -2,7 +2,7 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const { UserInputError } = require('apollo-server');
 
-const { validateRegisterInput } = require('../../util/validators');
+const { validateRegisterInput } = require('../../util/validatorRegisterInput');
 const { SECRET_KEY } = require('../../config');
 const User = require('../../models/User');
