+++ SayHiSemanticUi/util/validatorRegisterInput.js
